Refactore code in new pet form to use the same form for editing pet
