+++ client/src/components/EditPet.jsx
@@ -1,0 +1,10 @@
+import { NewPetForm } from "./NewPetForm"
+
+export const EditPet = () => {
+    return <div>
+        <div>Edit pet</div>
+        <div>
+            <NewPetForm/>
+        </div>
+    </div>
+}
