+++ client/src/components/NewPetForm.jsx
@@ -1,48 +1,9 @@
-import { useFormik } from "formik"
 import { SubmitFormBtn } from './SubmitFormBtn'
-import * as Yup from 'yup'
-import { createPet } from '../slices/pets/petApi'
-import { useDispatch } from 'react-redux'
-import { useNavigate } from "react-router-dom"
 import { Input } from "./Input"
 import { InputFile } from "./InputFile"
 import { Select } from "./Select"
 
-export const NewPetForm = () => {
-    const dispatch = useDispatch()
-    const navigate = useNavigate()
-    const initialValues = {
-        petName: "",
-        species: "",
-        breed: "",
-        age: "",
-        image: null
-    }
-
-    const validationSchema = Yup.object({
-        petName: Yup.string().required("Pet name is required "),
-        species: Yup.string().required(),
-        breed: Yup.string().required(),
-        age: Yup.string().required("Pet age is required "),
-        image: Yup.mixed().required()
-    })
-
-    const formik = useFormik({
-        initialValues,
-        validationSchema,
-        onSubmit: (data, { resetForm }) => {
-            const formData = new FormData()
-            formData.append('petName', data.petName)
-            formData.append('species', data.species)
-            formData.append('breed', data.breed)
-            formData.append('age', data.age)
-            formData.append('image', data.image)
-            dispatch(createPet(formData))
-            resetForm()
-            navigate("/user")
-        }
-    })
-
+export const NewPetForm = ({formik}) => {
     return <div>
         <form onSubmit={formik.handleSubmit} encType="multipart/form-data">
             <Input
